% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baymedr-package.R
\docType{package}
\name{baymedr-package}
\alias{baymedr-package}
\alias{baymedr}
\title{baymedr: Computation of Bayes Factors for Common Biomedical Designs}
\description{
\code{baymedr} provides functions for the computation of Bayes factors for
common biomedical research designs.
}
\details{
\tabular{ll}{ Package: \tab baymedr \cr Type: \tab Package \cr Version: \tab
0.1.1.9000 \cr License: \tab GPL-3 \cr Date: \tab 2022-09-01 \cr}
}
\seealso{
Useful links: \itemize{ \item
  \url{https://github.com/maxlinde/baymedr} \item Report bugs at
  \url{https://github.com/maxlinde/baymedr/issues}}
}
\author{
\strong{Maintainer}: Maximilian Linde --
  \email{maximilian.linde.92@gmail.com}

  Authors: \itemize{ \item Maximilian Linde (aut, cre) \item Don van
  Ravenzwaaij (aut) \item Jorge N. Tendeiro (aut)
  \item Quentin F. Gronau (ctb) }
}
