% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bf.R
\name{get_bf}
\alias{get_bf}
\title{Extract Bayes factor from S4 object}
\usage{
get_bf(object)
}
\arguments{
\item{object}{An S4 object of class 'baymedrSuperiority',
'baymedrEquivalence', or 'baymedrNonInferiority'.}
}
\value{
A numeric scalar, providing the Bayes factor from an S4 object.
}
\description{
This function extracts the Bayes factor from S4 objects, created from the
functions \code{\link{super_bf}}, \code{\link{equiv_bf}}, or
\code{\link{infer_bf}}.
}
\examples{
# Extract Bayes factor from \\code{\\link{super_bf}} using raw data:
mod_super <- super_bf(x = rnorm(100, 10, 15),
                      y = rnorm(130, 13, 10))

get_bf(object = mod_super)

# Extract Bayes factor from \\code{\\link{equiv_bf}} using raw data:
mod_equiv <- equiv_bf(x = rnorm(100, 10, 15),
                      y = rnorm(130, 13, 10))

get_bf(object = mod_equiv)

# Extract Bayes factor from \\code{\\link{infer_bf}} using raw data:
mod_infer <- infer_bf(x = rnorm(100, 10, 15),
                      y = rnorm(130, 13, 10),
                      ni_margin = 1)

get_bf(object = mod_infer)
}
