% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrand_t.R
\name{integrand_t}
\alias{integrand_t}
\title{Bla}
\usage{
integrand_t(delta, t, n, nu, mu_delta, gamma, kappa)
}
\arguments{
\item{delta}{A number}

\item{t}{A number}

\item{n}{A number}

\item{nu}{A number}

\item{mu_delta}{A number}

\item{gamma}{A number}

\item{kappa}{A number}
}
\value{
A number
}
\description{
Bla
}
\examples{
integrand_t(1, 2, 3, 4, 5, 6, 7)
}
