% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_show.R
\docType{methods}
\name{methods-show}
\alias{methods-show}
\alias{show,baymedrSuperiority-method}
\alias{show,baymedrEquivalence-method}
\alias{show,baymedrNonInferiority-method}
\title{Methods to print (show) S4 objects in the console}
\usage{
\S4method{show}{baymedrSuperiority}(object)

\S4method{show}{baymedrEquivalence}(object)

\S4method{show}{baymedrNonInferiority}(object)
}
\description{
Definition of costumised printing (showing) of the S4 objects
\linkS4class{baymedrSuperiority}, \linkS4class{baymedrEquivalence}, and
\linkS4class{baymedrNonInferiority}
}
