% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_bf.R
\name{super_bf}
\alias{super_bf}
\title{Bayes factors for superiority designs}
\usage{
super_bf(x = NULL, y = NULL, n_x = NULL, n_y = NULL,
  mean_x = NULL, mean_y = NULL, sd_x = NULL, sd_y = NULL,
  ci_margin = NULL, prior_scale = 1/sqrt(2))
}
\arguments{
\item{x}{A vector of numeric observations for the control group.}

\item{y}{A vector of numeric observations for the experimental group.}

\item{n_x}{A scalar, specifying the sample size of the control group.}

\item{n_y}{A scalar, specifying the sample size of the experimental group.}

\item{mean_x}{A scalar, specifying the mean of the dependent variable in the
control group.}

\item{mean_y}{A scalar, specifying the mean of the dependent variable in the
experimental group.}

\item{sd_x}{A scalar, specifying the standard deviation of the dependent
variable in the control group. Only sd_x and sd_y OR ci_margin should be
defined (see Details).}

\item{sd_y}{A scalar, specifying the standard deviation of the dependent
variable in the experimental group. Only sd_x and sd_y OR ci_margin should be
defined (see Details).}

\item{ci_margin}{A scalar, specifying the width of the confidence interval
of the difference on the dependent variable between the control and
experimental groups. Only sd_x and sd_y OR ci_margin should be defined
(see Details).}

\item{prior_scale}{A scalar, specifying the scale of the prior distribution
(see Details).}
}
\value{
Two Bayes factors are obtained from \code{super_bf}. The first one
  corresponds to a one-tailed alternative hypothesis (i.e., \eqn{\mu_control
  < \mu_experimental}), whereas the second one corresponds to a two-tailed
  alternative hypothesis. This is done to accomodate different research
  practices, with some researchers employing a one-tailed and others a
  two-tailed test. Importantly, both Bayes factors refer to the evidence in
  favour of the alternative hypothesis.
}
\description{
This function computes Bayes factors for superiority designs.
}
\details{
The Bayes factor resulting from \code{super_bf} tests the null hypothesis
that the experimental group (e.g., a new medication) is not better than the
control group (e.g., a placebo or existing medication). In other words, the
null hypothesis of \code{super_bf} states that the true population effect
size is exactly zero.

As the name 'superiority' implies, the test is typically understood as a
one-tailed test. In practice, however, a two-tailed test is often employed.
To cover both research practices, \code{super_bf} calculates two Bayes
factors, one for each alternative hypothesis (see Value).
}
\examples{
# super_bf using x and y:
super_bf(x = rnorm(100, 10, 15), y = rnorm(130, 13, 10))
}
\references{
Gronau, Q. F., Ly, A., & Wagenmakers, E.-J. (2018). Informed
bayesian t-tests. Manuscript submitted for publication.

Rouder, J. N., Speckman, P. L., Sun, D., & Morey, R. D. (2009). Bayesian t
tests for accepting and rejecting the null hypothesis. \emph{Psychonomic
Bulletin & Review}, \emph{16}(2), 225-237.

van Ravenzwaaij, D., Monden, R., Tendeiro, J. N., & Ioannidis, J. P. A.
(2019). Bayes factors for superiority, non-inferiority, and equivalence
designs. Manuscript submitted for publication.
}
