% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_bf.R
\name{infer_bf}
\alias{infer_bf}
\title{Bayes factor for non-inferiority designs}
\usage{
infer_bf(x = NULL, y = NULL, n_x = NULL, n_y = NULL,
  mean_x = NULL, mean_y = NULL, sd_x = NULL, sd_y = NULL,
  ni_margin = NULL, prior_scale = 1/sqrt(2))
}
\arguments{
\item{x}{A numeric vector of observations for the control group.}

\item{y}{A numeric vector of observations for the experimental group.}

\item{n_x}{A numeric scalar, specifying the sample size of the control group.}

\item{n_y}{A numeric scalar, specifying the sample size of the experimental
group.}

\item{mean_x}{A numeric scalar, specifying the mean of the dependent variable
in the control group.}

\item{mean_y}{A numeric scalar, specifying the mean of the dependent variable
in the experimental group.}

\item{sd_x}{A numeric scalar, specifying the standard deviation of the
dependent variable in the control group.}

\item{sd_y}{A numeric scalar, specifying the standard deviation of the
dependent variable in the experimental group.}

\item{ni_margin}{##TODO##}

\item{prior_scale}{A numeric scalar, specifying the scale of the prior
distribution (see Details). The default value is \eqn{1 / \sqrt{2}} (see
Rouder et al., 2009).}
}
\value{
##TODO##
}
\description{
This function computes a Bayes factor for non-inferiority designs.
}
\details{
The Bayes factor resulting from \code{\link{infer_bf}} tests the null
hypothesis that the experimental group (e.g., a new medication) is not better
than the control group (e.g., a placebo or existing medication) minus a
constant value (c). The alternative hypothesis is that the experimental group
is better than the control group minus a constant value (c). Put more
formally, the null hypothesis states that the true population effect size <
-c, resulting in the point null hypothesis that the true population effect
size = -c against the one-sided alternative hypothesis that the true
population effect size > -c.

Importantly, \code{\link{infer_bf}} can be utilized to calculate a Bayes
factor based on raw data (i.e., if arguments \code{x} and \code{y} are
defined) or summary statistics (i.e., if arguments \code{n_x}, \code{n_y},
\code{mean_x}, and \code{mean_y} are defined).

##TODO## ni_margin

##TODO## prior_scale
}
\examples{
# infer_bf using raw data:
infer_bf(x = rnorm(100, 10, 15),
         y = rnorm(130, 13, 10),
         ni_margin = -1)

# infer_bf using summary statistics:
infer_bf(n_x = 100,
         n_y = 130,
         mean_x = 10,
         mean_y = 13,
         sd_x = 15,
         sd_y = 10,
         ni_margin = -1)
}
\references{
Gronau, Q. F., Ly, A., & Wagenmakers, E.-J. (2018). Informed
  bayesian t-tests. Manuscript submitted for publication.

  Rouder, J. N., Speckman, P. L., Sun, D., & Morey, R. D. (2009). Bayesian t
  tests for accepting and rejecting the null hypothesis. \emph{Psychonomic
  Bulletin & Review}, \emph{16}(2), 225-237.

  van Ravenzwaaij, D., Monden, R., Tendeiro, J. N., & Ioannidis, J. P. A.
  (2019). Bayes factors for superiority, non-inferiority, and equivalence
  designs. Manuscript submitted for publication.
}
