% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_bf.R
\name{infer_bf}
\alias{infer_bf}
\title{Bayes factor for non-inferiority designs}
\usage{
infer_bf(x = NULL, y = NULL, n_x = NULL, n_y = NULL,
  mean_x = NULL, mean_y = NULL, sd_x = NULL, sd_y = NULL,
  ni_margin = NULL, prior_scale = 1/sqrt(2), alternative = "greater")
}
\arguments{
\item{x}{A numeric vector of observations for the control group.}

\item{y}{A numeric vector of observations for the experimental group.}

\item{n_x}{A numeric vector of length one, specifying the sample size of the
control group.}

\item{n_y}{A numeric vector of length one, specifying the sample size of the
experimental group.}

\item{mean_x}{A numeric vector of length one, specifying the mean of the
dependent variable in the control group.}

\item{mean_y}{A numeric vector of length one, specifying the mean of the
dependent variable in the experimental group.}

\item{sd_x}{A numeric vector of length one, specifying the standard deviation
of the dependent variable in the control group. Only \code{sd_x} and
\code{sd_y} \strong{OR} \code{ci_margin} should be defined (see Details).}

\item{sd_y}{A numeric vector of length one, specifying the standard deviation
of the dependent variable in the experimental group. Only \code{sd_x} and
\code{sd_y} \strong{OR} \code{ci_margin} should be defined (see Details).}

\item{ni_margin}{A numeric vector of length one, specifying the
non-inferiority margin in unstandardized units. The value should be a
positive number.}

\item{prior_scale}{A numeric vector of length one, specifying the scale of
the prior distribution (see Details). The default value is \eqn{1 /
\sqrt{2}}.}

\item{alternative}{A character vector of length one, specifying the direction
of the alternative hypothesis. 'greater' (the default) corresponds to mu_y
> mu_x  and 'less' refers to mu_y < mu_x (see Details).}
}
\value{
An S4 object of class 'baymedrSuperiority' is returned. Contained are
  a description of the model and the resulting Bayes factor: \itemize{ \item
  test: The type of analysis \item hypotheses: A statement of the hypotheses
  \itemize{ \item h0: The null hypothesis \item h1: The alternative
  hypothesis} \item ni_margin: The value for ni_margin \item data: A
  description of the data \itemize{ \item type: The type of data ('raw' when
  arguments \code{x} and \code{y} are used or 'summary' when arguments
  \code{n_x}, \code{n_y}, \code{mean_x}, \code{mean_y}, \code{sd_x}, and
  \code{sd_y} (or \code{ci_margin} instead of \code{sd_x} and \code{sd_y})
  are used) \item ...: values for the arguments used, depending on 'raw' or
  'summary'} \item prior_scale: The width of the Cauchy prior distribution
  \item bf: The resulting Bayes factor } A summary of the model is shown by
  printing the object.
}
\description{
This function computes a Bayes factor for non-inferiority designs.
}
\details{
The Bayes factor resulting from \code{\link{infer_bf}} tests the null
hypothesis that the experimental group (e.g., a new medication) is not better
than the control group (e.g., a placebo or existing medication) minus a
constant value (c). The alternative hypothesis is that the experimental group
is better than the control group minus a constant value (c). Put more
formally, the null hypothesis states that the true population effect size <
-c, resulting in the point null hypothesis that the true population effect
size = -c against the one-sided alternative hypothesis that the true
population effect size > -c.

Since sometimes high scores on the dependent variable are considered
non-inferior (e.g., amount of social interactions) and sometimes rather the
low scores (e.g., severity of symptoms), the user can specify the direction
of non-inferiority with the argument \code{alternative}. For the case where
higher values on the dependent variable indicate non-inferiority, the user
should specify 'greater' (the default) for the argument \code{alternative};
if lower values on the dependent variable indicate non-inferiority, 'less'
should be specified for the argument \code{alternative}.

Importantly, \code{\link{infer_bf}} can be utilized to calculate a Bayes
factor based on raw data (i.e., if arguments \code{x} and \code{y} are
defined) or summary statistics (i.e., if arguments \code{n_x}, \code{n_y},
\code{mean_x}, and \code{mean_y} are defined).

With the argument \code{ni_margin}, the user can determine the
non-inferiority margin in unstandardized units. \code{ni_margin} should be a
positive number. Following ethical scientific rigour, \code{ni_margin} should
be defined prior to data collection and data analysis.

The argument \code{prior_scale} specifies the width of the prior distribution
on effect size. This prior follows a Cauchy distribution with one degree of
freedom. Visually, it resembles a normal distribution, although with much
heavier tails (see e.g., Rouder et al., 2009). Mathematically, a Cauchy
distribution with one degree of freedom is equivalent to a normal
distribution with a mean of zero and a variance that follows an inverse
chi-square distribution with one degree of freedom, for which the variance is
integrated out (Liang et al., 2008). \code{prior_scale} corresponds to half
of the interquartile range of the Cauchy prior. In general, the larger the
value for \code{prior_scale}, the broader the Cauchy prior distribution, and
the higher the relative support for the null hypothesis, reflected in the
resulting Bayes factor.

\code{\link{infer_bf}} creates an S4 object of class 'baymedrNonInferiority',
which has multiple slots/entries (e.g., type of data, prior scale, Bayes
factor, etc.; see Value). If it is desired to store or extract solely the
Bayes factor, the user can do this with \code{\link{get_bf}}, by setting the
S4 object as an argument (see Examples).
}
\examples{
## infer_bf using raw data:

# Assign model to variable.
infer_raw <- infer_bf(x = rnorm(100, 10, 15),
                      y = rnorm(130, 13, 10),
                      ni_margin = -1)

# Extract Bayes factor from model.
get_bf(infer_raw)

# ----------
# ----------

## infer_bf using summary statistics with data from Andersson et al. (2013).
## Test at timepoint 1:

# Assign model to variable.
infer_sum_t1 <- infer_bf(n_x = 32,
                         n_y = 33,
                         mean_x = 13.6,
                         mean_y = 17.1,
                         sd_x = 9.8,
                         sd_y = 8,
                         ni_margin = 2)

# Extract Bayes factor from model
get_bf(infer_sum_t1)

# ----------
# ----------

## infer_bf using summary statistics with data from Andersson et al. (2013).
## Test at timepoint 2:

# Assign model to variable.
infer_sum_t2 <- infer_bf(n_x = 32,
                         n_y = 30,
                         mean_x = 9.2,
                         mean_y = 13.5,
                         sd_x = 7.6,
                         sd_y = 8.7,
                         ni_margin = 2)

# Extract Bayes factor from model
get_bf(infer_sum_t2)
}
\references{
Gronau, Q. F., Ly, A., & Wagenmakers, E.-J. (2019). Informed
  bayesian t-tests. \emph{The American Statistician}, 1-13.

  Liang, F., Paulo, R., Molina, G., Clyde, M. A., & Berger, J. O. (2008).
  Mixtures of g priors for bayesian variable selection. \emph{Journal of the
  American Statistical Association}, \emph{103}(481), 410-423.

  Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., & Iverson, G.
  (2009). Bayesian t tests for accepting and rejecting the null hypothesis.
  \emph{Psychonomic Bulletin & Review}, \emph{16}(2), 225-237.

  van Ravenzwaaij, D., Monden, R., Tendeiro, J. N., & Ioannidis, J. P. A.
  (2019). Bayes factors for superiority, non-inferiority, and equivalence
  designs. \emph{BMC Medical Research Methodology}, \emph{19}(1), 71.
}
