% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_bf.R
\name{super_bf}
\alias{super_bf}
\title{Bayes factors for superiority designs}
\usage{
super_bf(x = NULL, y = NULL, n_x = NULL, n_y = NULL,
  mean_x = NULL, mean_y = NULL, sd_x = NULL, sd_y = NULL,
  ci_margin = NULL, prior_scale = 1/sqrt(2))
}
\arguments{
\item{x}{A vector of numeric observations for the control group.}

\item{y}{A vector of numeric observations for the experimental group.}

\item{n_x}{A scalar, specifying the sample size of the control group.}

\item{n_y}{A scalar, specifying the sample size of the experimental group.}

\item{mean_x}{A scalar, specifying the mean of the dependent variable in the
control group.}

\item{mean_y}{A scalar, specifying the mean of the dependent variable in the
experimental group.}

\item{sd_x}{A scalar, specifying the standard deviation of the dependent
variable in the control group. Only \code{sd_x} and sd_y OR ci_margin should be
defined (see Details).}

\item{sd_y}{A scalar, specifying the standard deviation of the dependent
variable in the experimental group. Only sd_x and sd_y OR ci_margin should be
defined (see Details).}

\item{ci_margin}{A scalar, specifying the margin of the confidence interval
(i.e., the width of the confidence interval divided by 2) of the difference
on the dependent variable between the control and experimental groups. Only
sd_x and sd_y OR ci_margin should be defined (see Details).}

\item{prior_scale}{A scalar, specifying the scale of the prior distribution
(see Details). The default value is \eqn{1 / \sqrt{2}}
(see Rouder et al., 2009).}
}
\value{
Two Bayes factors are obtained from \code{super_bf}. The first one
  corresponds to a one-tailed alternative hypothesis (i.e., \eqn{\mu_control
  < \mu_experimental}), whereas the second one corresponds to a two-tailed
  alternative hypothesis. This is done to accomodate different research
  practices, with some researchers employing a one-tailed and others a
  two-tailed test. Importantly, both Bayes factors refer to the evidence in
  favour of the alternative hypothesis.
}
\description{
This function computes Bayes factors for superiority designs.
}
\details{
The Bayes factor resulting from \code{super_bf} tests the null hypothesis
that the experimental group (e.g., a new medication) is not better than the
control group (e.g., a placebo or existing medication). In other words, the
null hypothesis of \code{super_bf} states that the true population effect
size is exactly zero.

As the name 'superiority' implies, the test is typically understood as a
one-tailed test. In practice, however, a two-tailed test is often employed.
To cover both research practices, \code{super_bf} calculates two Bayes
factors, one for each alternative hypothesis (see Value).

Importantly, \code{super_bf} can be utilized to calculate Bayes factors
based on raw data (i.e., if arguments x and y are defined) or summary
statistics (i.e., if arguments n_x, n_y, mean_x, and mean_y are defined). In
the latter case, the user has the freedom to supply values either for the
arguments sd_x and sd_y OR ci_margin. The choice should depend on the
information that is available to the user.
}
\examples{
# super_bf using raw data:
super_bf(x = rnorm(100, 10, 15), y = rnorm(130, 13, 10))

# super_bf using summary statistics. The case where sd_x and sd_y are known:
super_bf(n_x = 100, n_y = 130,
         mean_x = 10, mean_y = 13,
         sd_x = 15, sd_y = 10)

# super_bf using summary statistics. The case where sd_x and sd_y are not
# known:
super_bf(n_x = 100, n_y = 130,
         mean_x = 10, mean_y = 13,
         ci_margin = 4)
}
\references{
Gronau, Q. F., Ly, A., & Wagenmakers, E.-J. (2018). Informed
bayesian t-tests. Manuscript submitted for publication.

Rouder, J. N., Speckman, P. L., Sun, D., & Morey, R. D. (2009). Bayesian t
tests for accepting and rejecting the null hypothesis. \emph{Psychonomic
Bulletin & Review}, \emph{16}(2), 225-237.

van Ravenzwaaij, D., Monden, R., Tendeiro, J. N., & Ioannidis, J. P. A.
(2019). Bayes factors for superiority, non-inferiority, and equivalence
designs. Manuscript submitted for publication.
}
